+++ src/main/java/uk/gov/digital/ho/pttg/application/retry/UnauthorizedHttpClientErrorExceptionRetryPolicy.java
@@ -1,4 +1,4 @@
-package uk.gov.digital.ho.pttg.api;
+package uk.gov.digital.ho.pttg.application.retry;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.retry.RetryContext;
